+++ script/InitDummyUsers.sol
@@ -16,17 +16,12 @@ import {ASC721Manager} from "../src/ASC721Manager.sol";
 
 contract InitDummyUsers is Script {
     ASC721Manager public asc;
-    address public constant ADMIN_ADDRESS =
-        0x98A5c7E6eb3DEaf7Db34d14d63D46ec5a5A2f775; // dummy CHANGE THIS FOR MAINNET
-    address public constant TREASURY_ADDRESS =
-        0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266; // dummy CHANGE THIS FOR MAINNET
+    address public constant ADMIN_ADDRESS = 0x98A5c7E6eb3DEaf7Db34d14d63D46ec5a5A2f775; // dummy CHANGE THIS FOR MAINNET
+    address public constant TREASURY_ADDRESS = 0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266; // dummy CHANGE THIS FOR MAINNET
 
-    address public constant ETH_FEE_PROXY_ADDRESS =
-        0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65; // base sepolia
-    address public constant LINK_ADDRESS =
-        0xE4aB69C077896252FAFBD49EFD26B5D171A32410; // base sepolia
-    address public constant VRF_WRAPPER_ADDRESS =
-        0x7a1BaC17Ccc5b313516C5E16fb24f7659aA5ebed; // base sepolia
+    address public constant ETH_FEE_PROXY_ADDRESS = 0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65; // base sepolia
+    address public constant LINK_ADDRESS = 0xE4aB69C077896252FAFBD49EFD26B5D171A32410; // base sepolia
+    address public constant VRF_WRAPPER_ADDRESS = 0x7a1BaC17Ccc5b313516C5E16fb24f7659aA5ebed; // base sepolia
 
     uint256 public constant ELEPHANT_ID = 0;
     uint256 public constant TIGER_ID = 1;
@@ -44,7 +39,7 @@ contract InitDummyUsers is Script {
         address dummy_1 = vm.envAddress("DUMMY_1_ADDRESS");
         address dummy_2 = vm.envAddress("DUMMY_2_ADDRESS");
         address[2] memory dummys = [dummy_1, dummy_2];
-        for (uint i = 0; i < dummys.length; i++) {
+        for (uint256 i = 0; i < dummys.length; i++) {
             asc.setKYC(dummys[i], true);
             asc.setEarlyBacker(dummys[i], true);
         }
